backend/requirements/base.txt
django==2.1.2
django-environ==0.4.5
django-extensions==2.1.2
djangorestframework==3.8.2
django-cors-headers==2.4.0
django-ipware==2.1.0
pillow==5.2.0
djangorestframework-camel-case==0.2.0
psycopg2==2.7.3.2
djangorestframework-jwt==1.11.0
django-guardian==1.4.9
structlog==18.2.0
ipython==7.1.1
django-extra-fields==1.0.0
